sing up and login routes added with authorization controllers. Refactored code to use async/await instead then blocks

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 
 const { validationResult } = require('express-validator/check');
 const Post = require('../models/post');
+const User = require('../models/user');
 
 
 const deleteImage =  filePath => {
@@ -10,20 +11,32 @@ const deleteImage =  filePath => {
     fs.unlink(filePath, err => console.log(err));
   };
 
-exports.getPosts = (req, res, next) => {
-    Post.find()
-        .then(posts => {
-            res.status(200).json({ posts: posts })
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
-        })
+exports.getPosts = async (req, res, next) => {
+    const page = req.query.page || 1;
+    const perPage = 2;
+    
+    try {
+        const totalItems = await Post.find().countDocuments();            
+        const posts = await Post.find()
+            .populate('creator')
+            .skip((page - 1) * perPage)
+            .limit(perPage);
+        
+        res.status(200)
+            .json({ 
+                posts: posts, 
+                totalItems: totalItems 
+            })
+
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }        
 }
 
-exports.createPost = (req, res, next) => {
+exports.createPost = async (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty) {
         const error = new Error('Validation failed, entered data is incorrect');
@@ -40,52 +53,54 @@ exports.createPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     const imageUrl = req.file.path.replace('\\','/');
-    console.log(imageUrl);
-    const post = new Post({
+    const newPost = new Post({
             title: title,
             imageUrl: imageUrl,
             content: content,
-            creator: {
-                name: 'Damian'
-            }
+            creator: req.userId
         });
-    post.save()
-        .then(result => {
-            res.status(201).json({
-                message: 'Post created successfully!',
-                post: result
-                })
+    newPost.save()
+    try {
+        const user = await User.findById(req.userId);
+        user.posts.push(newPost);
+        user.save();
+        const post = await Post.findOne(newPost._id)
+            .populate('creator', 'name');
+        res.status(201).json({
+            message: 'Post created successfully!',
+            post: post,
+            creator: { _id: user._id, name: user.name }
         })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
-        })
+    } catch (err) {
+        if (!err.statusCode) {
+        err.statusCode = 500;
+        }
+        next(err);
+    }
 }
 
-exports.getPost = (req, res, next) => {
+exports.getPost = async (req, res, next) => {
     const postId = req.params.postId;
-    Post.findById(postId)
-        .then(post => {
-            if (!post) {
-                const error = new Error('Post was not found');
-                error.statusCode = 404;
-                throw error;
-            }
-            res.status(200).json({
-                post: post
-            })
+    const post = await Post.findById(postId);
+    try {
+        if (!post) {
+            const error = new Error('Post was not found');
+            error.statusCode = 404;
+            throw error;
+        }
+        res.status(200).json({
+            post: post
         })
-        .catch(err => {
+
+    } catch(err) {
             if (!err.statusCode) {
                 err.statusCode = 500;
             }
             next(err);
-        })
+    }
 }
 
-exports.updatePost = (req, res, next) => {
+exports.updatePost = async (req, res, next) => {
     const postId = req.params.postId;
     const errors = validationResult(req);
     if (!errors.isEmpty) {
@@ -107,29 +122,61 @@ exports.updatePost = (req, res, next) => {
         throw error;
     }
 
-    Post.findById(postId)
-        .then(post => {
-            if (!post) {
-                const error = new Error('Post was not found');
-                error.statusCode = 404;
-                throw error;
-            }
-            if ( updatedImageUrl !== post.imageUrl ) {
-                deleteImage(post.imageUrl);
-            }
-            post.title = updatedTitle;
-            post.imageUrl = updatedImageUrl;
-            post.content = updatedContent;
-            return post.save();
-        })
-        .then(result => {
-            res.status(200).json({ post: result });
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
-        })
+    const post = await Post.findById(postId);
+    try {
+        if (!post) {
+            const error = new Error('Post was not found');
+            error.statusCode = 404;
+            throw error;
+        }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error;
+        }
+        if ( updatedImageUrl !== post.imageUrl ) {
+            deleteImage(post.imageUrl);
+        }
+        post.title = updatedTitle;
+        post.imageUrl = updatedImageUrl;
+        post.content = updatedContent;
+        post.save();
+        res.status(200).json({ post: post });
+    } catch(err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
 }
 
+exports.deletePost = async (req, res, next) => {
+    const postId = req.params.postId;
+    const post = await Post.findById(postId)
+    try {
+        if (!post) {
+            const error = new Error('Post was not found');
+            error.statusCode = 404;
+            throw error;
+        };
+
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error;
+        };
+
+        deleteImage(post.imageUrl);
+        await Post.findByIdAndDelete(postId);
+
+        const user = await User.findById(req.userId);
+        user.posts.pull(postId)
+        user.save();
+        res.status(200).json({ message: 'Post deleted' })
+    } catch(err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
+}
